Refact: Refactoring for the tabs switch on about page
